Change cost type in hkaiStreamingSetGraphConnection

+++ HKX2/Manual/hkaiStreamingSetGraphConnection.cs
@@ -2,7 +2,7 @@ namespace HKX2
 {
     public class hkaiStreamingSetGraphConnection : IHavokObject
     {
-        public short m_edgeCost;
+        public System.Half m_edgeCost;
         public uint m_edgeData;
 
         public int m_nodeIndex;
@@ -14,7 +14,7 @@ namespace HKX2
             m_nodeIndex = br.ReadInt32();
             m_oppositeNodeIndex = br.ReadInt32();
             m_edgeData = br.ReadUInt32();
-            m_edgeCost = br.ReadInt16();
+            m_edgeCost = br.ReadHalf();
             br.ReadUInt16();
         }
 
@@ -23,7 +23,7 @@ namespace HKX2
             bw.WriteInt32(m_nodeIndex);
             bw.WriteInt32(m_oppositeNodeIndex);
             bw.WriteUInt32(m_edgeData);
-            bw.WriteInt16(m_edgeCost);
+            bw.WriteHalf(m_edgeCost);
             bw.WriteUInt16(0);
         }
     }
